+++ python/test/test_imagesloader.py
@@ -3,7 +3,7 @@ import os
 import unittest
 from nose.tools import assert_equals, assert_true, assert_almost_equal
 
-from thunder.rdds.rddio.imagesloader import ImagesLoader
+from thunder.rdds.fileio.imagesloader import ImagesLoader
 from test_images import _have_image
 from test_utils import PySparkTestCase
