+++ python/thunder/rdds/fileio/writers.py
@@ -3,7 +3,7 @@ import shutil
 import urllib
 import urlparse
 
-from thunder.rdds.rddio.readers import _BotoS3Client, getByScheme
+from thunder.rdds.fileio.readers import _BotoS3Client, getByScheme
 
 
 _have_boto = False
