+++ python/thunder/utils/context.py
@@ -3,8 +3,8 @@
 import os
 import json
 from numpy import asarray, floor, ceil
-from thunder.rdds.rddio.imagesloader import ImagesLoader
-from thunder.rdds.rddio.seriesloader import SeriesLoader
+from thunder.rdds.fileio.imagesloader import ImagesLoader
+from thunder.rdds.fileio.seriesloader import SeriesLoader
 from thunder.utils.datasets import DataSets
 from thunder.utils.common import checkparams
